Feature: Check Declaration In Previous Scope
When declaring in scopes, it is important to know that if you have a declaration and the same declaration inside a if, it is wrong. Or having a int declaration and then creating a for with the same declaration.

For that it is important to check in previous.

Code check in previous scope that it is not the current nor the global scope to differ between global and local.

* Includes Fixes for method naming